Consolidate code for course selection: only single form needed now that selection aided by autocomplete
