+++ catalogue_app/templates/includes/_navbar.html
@@ -15,7 +15,7 @@
 				<li class="dropdown li-border-right">
 					<a href="" class="dropdown-toggle" data-toggle="dropdown">{{ _('Catalogue') }}<span class="caret"></span></a>
 					<ul class="dropdown-menu">
-						<li><a href="{{ url_for('course.course_code_selection') }}">{{ _('Search by Course') }}</a></li>
+						<li><a href="{{ url_for('course.course_selection') }}">{{ _('Search by Course') }}</a></li>
 						<li><a href="{{ url_for('course.departmental') }}">{{ _('Search by Department') }}</a></li>
 					</ul>
 				</li>
